requirements.txt
arviz==0.11.2
asgiref==3.3.1
backcall==0.2.0
backports.csv==1.0.7
certifi==2020.6.20
cffi==1.14.5
cftime==1.4.1
chardet==4.0.0
colorama==0.4.3
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
django==3.1.7
django-allauth==0.44.0
django-bootstrap-icons==0.2.5
django-crispy-forms==1.11.2
ephem==3.7.7.1
gunicorn==20.1.0
hijri-converter==2.1.1
holidays==0.10.5.2
idna==2.10
iexfinance==0.5.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.6.3
matplotlib==3.4.0
multitasking==0.0.9
netCDF4==1.5.6
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.1
PyJWT==2.0.1
PyMeeus==0.5.10
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pyzmq==19.0.2
requests==2.25.1
requests-cache==0.5.2
requests-oauthlib==1.3.0
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tornado==6.0.4
tqdm==4.59.0
traitlets==5.0.4
urllib3==1.26.4
wcwidth==0.2.5
wincertstore==0.2
xarray==0.17.0
Yahoo-ticker-downloader==3.0.1
yfinance==0.1.59
stocknews==0.9.11
scikit-learn==0.24.1
xmltodict==0.12.0
statsmodels==0.12.2
pmdarima==1.8.2